requirements.txt
psycopg
pandas
requests
python-dotenv
yfinance